feature(refactor): Reorganize the data structure
