+++ src/engines/mnt6_753/mod.rs
@@ -1,12 +1,12 @@
 mod ec;
 mod fq;
-mod fq2;
-mod fq4;
+mod fq3;
+mod fq6;
 mod fr;
 
 use self::{
     ec::g2::{ AteDoubleCoefficients, AteAdditionCoefficients, G2ProjectiveExtended },
-    fq::{ MNT4_X, MNT4_X_IS_NEGATIVE, TWIST_INV, TWIST, EXP_W0, EXP_W1, EXP_W0_IS_NEGATIVE }
+    fq::{ MNT6_X, MNT6_X_IS_NEGATIVE, TWIST_INV, TWIST, EXP_W0, EXP_W1, EXP_W0_IS_NEGATIVE }
 };
 
 pub use self::ec::{
@@ -15,29 +15,29 @@ pub use self::ec::{
 };
 
 pub use self::fq::{Fq, FqRepr};
-pub use self::fq2::Fq2;
-pub use self::fq4::Fq4;
+pub use self::fq3::Fq3;
+pub use self::fq6::Fq6;
 pub use self::fr::{Fr, FrRepr};
 
-use super::{Engine, CurveAffine};
+use crate::{Engine, CurveAffine};
 
 use ff::{BitIterator, Field, ScalarEngine};
 
 #[derive(Clone, Debug)]
-pub struct Mnt4;
+pub struct Mnt6;
 
-impl Mnt4 {
+impl Mnt6 {
 
     fn ate_pairing_loop(
         p: &G1Prepared,
         q: &G2Prepared,
-    ) -> Fq4 {
+    ) -> Fq6 {
 
-        let mut l1_coeff = Fq2::zero();
+        let mut l1_coeff = Fq3::zero();
         l1_coeff.c0 = p.p.x;
         l1_coeff.sub_assign(&q.x_over_twist);
 
-        let mut f = Fq4::one();
+        let mut f = Fq6::one();
 
         let mut dbl_idx: usize = 0;
         let mut add_idx: usize = 0;
@@ -45,7 +45,7 @@ impl Mnt4 {
 
         // The for loop is executed for all bits (EXCEPT the MSB itself) of
         let mut found_one = false;
-        for bit in BitIterator::new(&MNT4_X).skip(1) {
+        for bit in BitIterator::new(&MNT6_X).skip(1) {
             if !found_one {
                 found_one = bit;
                 continue;
@@ -54,7 +54,7 @@ impl Mnt4 {
             let dc = &q.double_coefficients[dbl_idx];
             dbl_idx += 1;
 
-            let mut g_rr_at_p = Fq4::zero();
+            let mut g_rr_at_p = Fq6::zero();
 
             let mut t0 = dc.c_j;
             t0.mul_assign(&p.x_by_twist);
@@ -75,7 +75,7 @@ impl Mnt4 {
                 let ac = &q.addition_coefficients[add_idx];
                 add_idx += 1;
 
-                let mut g_rq_at_p = Fq4::zero();
+                let mut g_rq_at_p = Fq6::zero();
 
                 let mut t0 = ac.c_rz;
                 t0.mul_assign(&p.y_by_twist);
@@ -95,10 +95,10 @@ impl Mnt4 {
             }
         }
 
-        if MNT4_X_IS_NEGATIVE {
+        if MNT6_X_IS_NEGATIVE {
             let ac = &q.addition_coefficients[add_idx];
 
-            let mut g_rnegr_at_p = Fq4::zero();
+            let mut g_rnegr_at_p = Fq6::zero();
 
             let mut t0 = ac.c_rz;
             t0.mul_assign(&p.y_by_twist);
@@ -121,27 +121,32 @@ impl Mnt4 {
         f
     }
 
-    fn final_exponentiation_part_one(elt: &Fq4, elt_inv: &Fq4) -> Fq4 {
-        /* (q^2-1) */
+    fn final_exponentiation_part_one(elt: &Fq6, elt_inv: &Fq6) -> Fq6 {
+        // (q^3-1)*(q+1)
 
-        /* elt_q2 = elt^(q^2) */
-        let mut elt_q2 = elt.clone();
-        elt_q2.frobenius_map(2);
-        /* elt_q2_over_elt = elt^(q^2-1) */
-        let mut elt_q2_over_elt = elt_q2;
-        elt_q2_over_elt.mul_assign(&elt_inv);
+        // elt_q3 = elt^(q^3)
+        let mut elt_q3 = elt.clone();
+        elt_q3.frobenius_map(3);
+        // elt_q3_over_elt = elt^(q^3-1)
+        let mut elt_q3_over_elt = elt_q3;
+        elt_q3_over_elt.mul_assign(&elt_inv);
+        // alpha = elt^((q^3-1) * q)
+        let mut alpha = elt_q3_over_elt.clone();
+        alpha.frobenius_map(1);
+        // beta = elt^((q^3-1)*(q+1)
+        alpha.mul_assign(&elt_q3_over_elt);
 
-        elt_q2_over_elt
+        alpha
     }
 
-    fn final_exponentiation_part_two(elt: &Fq4, &elt_inv: &Fq4) -> Fq4 {
+    fn final_exponentiation_part_two(elt: &Fq6, &elt_inv: &Fq6) -> Fq6 {
         let mut elt_q = elt.clone();
         elt_q.frobenius_map(1);
 
-        let mut w1_part = elt_q.clone().cyclotomic_exp(&EXP_W1);
+        let mut w1_part = elt_q.cyclotomic_exp(&EXP_W1);
         let w0_part = match EXP_W0_IS_NEGATIVE {
-            true => elt_inv.clone().cyclotomic_exp(&EXP_W0),
-            false => elt.clone().cyclotomic_exp(&EXP_W0),
+            true => elt_inv.cyclotomic_exp(&EXP_W0),
+            false => elt.cyclotomic_exp(&EXP_W0),
         };
 
         w1_part.mul_assign(&w0_part);
@@ -150,18 +155,18 @@ impl Mnt4 {
 
 }
 
-impl ScalarEngine for Mnt4 {
+impl ScalarEngine for Mnt6 {
     type Fr = Fr;
 }
 
-impl Engine for Mnt4 {
+impl Engine for Mnt6 {
     type G1 = G1;
     type G1Affine = G1Affine;
     type G2 = G2;
     type G2Affine = G2Affine;
     type Fq = Fq;
-    type Fqe = Fq2;
-    type Fqk = Fq4;
+    type Fqe = Fq3;
+    type Fqk = Fq6;
 
     fn miller_loop<'a, I>(i: I) -> Self::Fqk
     where
@@ -172,7 +177,7 @@ impl Engine for Mnt4 {
             ),
         >,
     {
-        let mut f = Fq4::one();
+        let mut f = Fq6::one();
         for (p, q) in i.into_iter() {
             if !p.is_zero() && !q.is_zero() {
                 f.mul_assign(&Self::ate_pairing_loop(p, q));
@@ -182,7 +187,7 @@ impl Engine for Mnt4 {
         f
     }
 
-    fn final_exponentiation(f: &Fq4) -> Option<Fq4> {
+    fn final_exponentiation(f: &Fq6) -> Option<Fq6> {
         let value_inv = f.inverse();
         if value_inv.is_none() {
             return None;
@@ -203,8 +208,8 @@ impl G2Prepared {
     pub fn from_affine(q: G2Affine) -> Self {
         let mut res = G2Prepared {
             p: q,
-            x_over_twist: Fq2::zero(),
-            y_over_twist: Fq2::zero(),
+            x_over_twist: Fq3::zero(),
+            y_over_twist: Fq3::zero(),
             double_coefficients:   vec![],
             addition_coefficients: vec![],
         };
@@ -213,77 +218,77 @@ impl G2Prepared {
     }
 
     fn doubling_step(r: &mut G2ProjectiveExtended) -> AteDoubleCoefficients {
-        let mut a = r.t.clone();
+        let mut a = r.t;
         a.square();
-        let mut b = r.x.clone();
+        let mut b = r.x;
         b.square();
-        let mut c = r.y.clone();
+        let mut c = r.y;
         c.square();
-        let mut d = c.clone();
+        let mut d = c;
         d.square();
 
-        let mut e = r.x.clone();
+        let mut e = r.x;
         e.add_assign(&c);
         e.square();
         e.sub_assign(&b);
         e.sub_assign(&d);
 
-        let mut f = G2::get_coeff_a();
+        let mut f = G2Affine::get_coeff_a();
         f.mul_assign(&a);
         f.add_assign(&b);
         f.add_assign(&b);
         f.add_assign(&b);
 
-        let mut g = f.clone();
+        let mut g = f;
         g.square();
 
-        let mut d_eight = d.clone();
+        let mut d_eight = d;
         d_eight.double();
         d_eight.double();
         d_eight.double();
 
-        let mut t0 = e.clone();
+        let mut t0 = e;
         t0.double();
         t0.double();
 
-        let mut x = g.clone();
+        let mut x = g;
         x.sub_assign(&t0);
 
-        let mut y = e.clone();
+        let mut y = e;
         y.double();
         y.sub_assign(&x);
         y.mul_assign(&f);
         y.sub_assign(&d_eight);
 
-        let mut t0 = r.z.clone();
+        let mut t0 = r.z;
         t0.square();
 
-        let mut z = r.y.clone();
+        let mut z = r.y;
         z.add_assign(&r.z);
         z.square();
         z.sub_assign(&c);
         z.sub_assign(&t0);
 
-        let mut t = z.clone();
+        let mut t = z;
         t.square();
 
-        let mut c_h = z.clone();
+        let mut c_h = z;
         c_h.add_assign(&r.t);
         c_h.square();
         c_h.sub_assign(&t);
         c_h.sub_assign(&a);
 
-        let mut c_4c = c.clone();
+        let mut c_4c = c;
         c_4c.double();
         c_4c.double();
 
-        let mut c_j = f.clone();
+        let mut c_j = f;
         c_j.add_assign(&r.t);
         c_j.square();
         c_j.sub_assign(&g);
         c_j.sub_assign(&a);
 
-        let mut c_l = f.clone();
+        let mut c_l = f;
         c_l.add_assign(&r.x);
         c_l.square();
         c_l.sub_assign(&g);
@@ -300,68 +305,68 @@ impl G2Prepared {
     }
 
     fn addition_step(
-        x: &Fq2,
-        y: &Fq2,
+        x: &Fq3,
+        y: &Fq3,
         r: &mut G2ProjectiveExtended)
     -> AteAdditionCoefficients {
 
         let mut a = y.clone();
         a.square();
-        let mut b = r.t.clone();
+        let mut b = r.t;
         b.mul_assign(&x);
 
-        let mut d = r.z.clone();
+        let mut d = r.z;
         d.add_assign(&y);
         d.square();
         d.sub_assign(&a);
         d.sub_assign(&r.t);
         d.mul_assign(&r.t);
 
-        let mut h = b.clone();
+        let mut h = b;
         h.sub_assign(&r.x);
 
-        let mut i = h.clone();
+        let mut i = h;
         i.square();
 
-        let mut e = i.clone();
+        let mut e = i;
         e.double();
         e.double();
 
-        let mut j = h.clone();
+        let mut j = h;
         j.mul_assign(&e);
 
-        let mut v = r.x.clone();
+        let mut v = r.x;
         v.mul_assign(&e);
 
-        let mut l1 = d.clone();
+        let mut l1 = d;
         l1.sub_assign(&r.y);
         l1.sub_assign(&r.y);
 
-        let mut x = l1.clone();
+        let mut x = l1;
         x.square();
         x.sub_assign(&j);
         x.sub_assign(&v);
         x.sub_assign(&v);
 
-        let mut t0 = r.y.clone();
+        let mut t0 = r.y;
         t0.double();
         t0.mul_assign(&j);
 
-        let mut y = v.clone();
+        let mut y = v;
         y.sub_assign(&x);
         y.mul_assign(&l1);
         y.sub_assign(&t0);
 
-        let mut z = r.z.clone();
+        let mut z = r.z;
         z.add_assign(&h);
         z.square();
         z.sub_assign(&r.t);
         z.sub_assign(&i);
 
-        let mut t = z.clone();
+        let mut t = z;
         t.square();
 
-        let coeff = AteAdditionCoefficients { c_l1: l1, c_rz: z.clone() };
+        let coeff = AteAdditionCoefficients { c_l1: l1, c_rz: z };
 
         r.x = x;
         r.y = y;
@@ -388,12 +393,12 @@ impl G2Prepared {
         let mut r = G2ProjectiveExtended {
             x: self.p.x,
             y: self.p.y,
-            z: Fq2::one(),
-            t: Fq2::one(),
+            z: Fq3::one(),
+            t: Fq3::one(),
         };
 
         let mut found_one = false;
-        for bit in BitIterator::new(&MNT4_X).skip(1) {
+        for bit in BitIterator::new(&MNT6_X).skip(1) {
 
             if !found_one {
                 found_one = bit;
@@ -409,7 +414,7 @@ impl G2Prepared {
             }
         }
 
-        if MNT4_X_IS_NEGATIVE {
+        if MNT6_X_IS_NEGATIVE {
             let rz_inv = r.z.inverse().expect("z should not be equal to zero");
             let mut rz2_inv = rz_inv;
             rz2_inv.square();
@@ -440,6 +445,7 @@ impl G1Prepared {
     }
 
     pub fn from_affine(p: G1Affine) -> Self {
+
         let mut res = G1Prepared {
             p: p,
             x_by_twist: TWIST,
@@ -457,6 +463,6 @@ impl G1Prepared {
 }
 
 #[test]
-fn mnt4_engine_tests() {
-    crate::tests::engine::engine_tests::<Mnt4>();
+fn mnt6_engine_tests() {
+    crate::tests::engine::engine_tests::<Mnt6>();
 }
